backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
channels==4.0.0
channels_redis==4.2.0
# Django==4.2.5
django-ledger==0.5.6.2

Django
django-filter==23.3
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
hl7==0.4.5
inflection==0.5.1

impacket

jsonschema==4.19.1
jsonschema-specifications==2023.7.1
PyJWT==2.8.0
pyserial==3.5
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4

pysmb==1.2.9.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
chardet==5.2.0
pillow==10.1.0
reportlab==4.0.8
celery==5.3.6
redis==5.0.1

uvicorn==0.27.1
uvicorn[standard]

weasyprint==60.2